refactor: migrate server.js to typescript
